requirements.txt
# requirements.txt
tree-sitter==0.20.1
tree-sitter-languages==1.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.28.1
click==8.1.7
numpy==1.24.3
astroid==2.15.5